Move claude agent into cluade_agent dir
