+++ agents/templates/claude_agent/__init__.py
@@ -1,0 +1,5 @@
+from .claude_agents import ClaudeCodeAgent
+from .claude_recorder import ClaudeCodeRecorder
+from .claude_tools import create_arc_tools_server
+
+__all__ = ['ClaudeCodeAgent', 'ClaudeCodeRecorder', 'create_arc_tools_server']
